requirements.txt
ansitable==0.11.4
cfgv==3.4.0
charset-normalizer==3.4.2
colored==2.3.0
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
fonttools==4.57.0
identify==2.6.10
kiwisolver==1.4.8
matplotlib==3.10.1
nodeenv==1.9.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
packaging==25.0
pgraph-python==0.6.3
pillow==11.2.1
platformdirs==4.3.7
pre-commit==4.2.0
progress==1.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
reportlab==4.4.2
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.15.2
six==1.17.0
spatialgeometry==1.1.0
spatialmath-python==1.1.14
swift-sim==1.1.0
typing-extensions==4.13.2
virtualenv==20.30.0
websockets==15.0.1
